Uni as funcionalidades de cadastrar e editar em um único Form
