+++ Cod3rsGrowth.Forms/Cod3rsGrowth.Forms.csproj.user
@@ -4,19 +4,13 @@
     <Compile Update="FormPrincipal.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Update="Forms\FormularioCadastroPersonagem.cs">
-      <SubType>Form</SubType>
-    </Compile>
     <Compile Update="Forms\FormFiltrar.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Update="Forms\FormularioCadastroHabilidade.cs">
+    <Compile Update="Forms\FormularioHabilidade.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Update="Forms\FormularioEditarHabilidade.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Update="Forms\FormularioEditarPersonagem.cs">
+    <Compile Update="Forms\FormularioPersonagem.cs">
       <SubType>Form</SubType>
     </Compile>
   </ItemGroup>
